refactor(core): move storage to separate file
